Implement GraphQL functionality for book-service and update front-end.

+++ book-service/src/main/java/com/chrisbaileydeveloper/bookservice/util/DataLoader.java
@@ -21,27 +21,27 @@ public class DataLoader implements CommandLineRunner {
             Book book1 = new Book();
             book1.setName("Design Patterns: Elements of Reusable Object-Oriented Software");
             book1.setDescription("A classic book on software design patterns for object-oriented programming.");
-            book1.setPrice(BigDecimal.valueOf(29));
+            book1.setPrice(29.0);
 
             Book book2 = new Book();
             book2.setName("The Pragmatic Programmer: Your Journey to Mastery");
             book2.setDescription("A comprehensive guide to the art of computer programming and best practices.");
-            book2.setPrice(BigDecimal.valueOf(42));
+            book2.setPrice(42.0);
 
             Book book3 = new Book();
             book3.setName("Clean Code: A Handbook of Agile Software Craftsmanship");
             book3.setDescription("A book focused on writing clean, maintainable code.");
-            book3.setPrice(BigDecimal.valueOf(35));
+            book3.setPrice(35.0);
 
             Book book4 = new Book();
             book4.setName("Refactoring: Improving the Design of Existing Code");
             book4.setDescription("A guide to refactoring and improving existing code.");
-            book4.setPrice(BigDecimal.valueOf(40));
+            book4.setPrice(40.0);
 
             Book book5 = new Book();
             book5.setName("Code Complete: A Practical Handbook of Software Construction");
             book5.setDescription("A comprehensive guide to software construction and best practices.");
-            book5.setPrice(BigDecimal.valueOf(45));
+            book5.setPrice(45.0);
 
             bookRepository.save(book1);
             bookRepository.save(book2);
